+++ NEWS.md
@@ -58,4 +58,5 @@
 * Added binomial distribution
 * Added ability to specify formula in variance
 * Added function genMultiFac - generates multi-factorial design data
+* Added function addMultiFac - adds multi-factorial design data